hypermax search algorithm with TT(move ordering and PV extract only), mvv-lva, killer table and history table

+++ engine/src/main/kotlin/pl/tomaszstankowski/fourplayerchess/engine/Api.kt
@@ -222,7 +222,7 @@ data class UIState(
         val legalMoves: Set<Move>
 )
 
-data class StateEvaluation(val principalVariation: List<PVMove>, val value: Int) {
+data class StateEvaluation(val principalVariation: List<PVMove>, val value: Float) {
     data class PVMove(
             val move: Move,
             val moveText: String)
